+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - SingleGradationDecal のマテリアル選択候補の表示に Missing や Clone のマテリアルが表示される問題を修正 (#563)
+- GameObject TexTransTool のメニューで生成される GameObject が常にワールド原点に生成されていた問題を修正 (#)
+- GameObject TexTransTool のメニューで生成される GameObject が自動で選択されない仕様を修正 (#)
 
 ## [v0.7.6](https://github.com/ReinaS-64892/TexTransTool/compare/v0.7.5...v0.7.6) - 2024-07-14